[FIX] payment: accept invoice ids in payment links
This commit brings back the possibility to pass invoice ids in payment
links, a feature that was lost with commit 5639960a. If such an id is
present in the payment link URL, the transaction that is created must
have a reference starting with the name of the invoice, and must be
linked to it.

task-2494916

closes odoo/odoo#72992

X-original-commit: a281c30289906f4c3d3dffe717ad8235f46c7dbc
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>