[FIX] mass_mailing: Fix crash when trying to schedule a new compaing from the calendar view
To reproduce the error:

1. Open Email marketing
2. Go to the calendar view
3. Click on a cell

Explanation:

The function 'default_get' defines an argument named 'field' that conflicts
with the 'field' variable imported from the default 'odoo' package. When
two identifiers conflict, Python will use the variable defined in the closest
enclosing scope.

When the 'default_get' function accesses the 'Datetime' attribute of 'field',
the function will raise an exception as the identifier 'field' will refer to
the argument of the function and the property 'Datetime' will not defined on
that variable.

To solve the issue, we will rename the argument of the function 'default_get'
to 'fields_list' so that it will no longer conflict with the imported variable
'fields'.

Links:

Task id: 2581357

closes odoo/odoo#72857

X-original-commit: e2ffc806123cb42c95fdcd429323866c9e8d66fe
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>